requirements.txt
0.0.1==0.0.1
3==0.0.0
4==8.8.8.8
absl-py==1.0.0
asgiref==3.5.0
astunparse==1.6.3
beautifulsoup4==4.10.0
bleach==4.1.0
bs4==0.0.1
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.11
django-allauth==0.47.0
django-appconf==1.0.5
django-bootstrap4==21.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-imagekit==4.1.0
django-markdownx==3.0.1
django-summernote==0.8.20.0
easy-thumbnails==2.8.1
flatbuffers==2.0
gast==0.4.0
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.1
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
lxml==4.7.1
Markdown==3.3.6
numpy==1.21.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pilkit==2.0
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.6
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
svglib==1.2.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tinycss2==1.1.1
typing_extensions==4.0.1
urllib3==1.26.8
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
